feat: Add binary label to training data
This commit adds a new feature to the data preparation process
by creating a binary label ('toxic_binary') indicating whether
a text is toxic or not.

The 'toxic_binary' column is created by summing the values of
the existing toxic labels (toxic, severe_toxic, obscene, threat,
insult, identity_hate) for each text. If the sum is greater than
0, the text is considered toxic and assigned a value of 1;
otherwise, it is considered non-toxic and assigned a value of 0.

This binary label simplifies the modeling task and allows for
training a binary classification model.