feat: Add binary label creation to DataPreparation
Adds a binary label creation function to the `DataPreparation` class.
This function takes a list of toxicity columns and adds a new column
"is_toxic" that indicates whether a comment is toxic based on any of
the specified toxicity labels. This change is necessary to simplify
toxicity classification and allow for more flexible toxicity
detection based on various labels.